requirements.txt
phonenumbers
beautifulsoup
requests
python-levenshtein
